pyproject.toml
[build-system]
requires = ["setuptools>=82.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langcore-audit"
version = "1.1.4"
description = "LangCore audit-logging provider — decorator that wraps any provider with structured audit trails"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "langcore>=1.0.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
]

[project.entry-points."langcore.providers"]
audit = "langcore_audit.provider:AuditLanguageModel"

[tool.setuptools.packages.find]
where = ["."]
include = ["langcore_audit*"]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pylint.main]
# langcore exposes `providers` dynamically; allow pylint to introspect it.
extension-pkg-allow-list = ["langcore"]

[tool.pylint."messages control"]
disable = ["no-member"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["langcore_audit"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["langcore_audit/__init__.py:__version__"]
branch = "main"
build_command = "pip install build && python -m build"
commit_message = "chore(release): v{version}"
tag_format = "v{version}"

[tool.semantic_release.changelog]
generate = true

[tool.semantic_release.publish]
upload_to_pypi = true
